chore: move measurement checking to libs/test-utils
Also rename battery-test-utils to battery-fakes to reflect changes.

This allows using the same measurement helpers for other app tests.
